fix: check breadcrumb is not empty before changing last child style and enable word break in breadcrumb display
In some cases (notification page), it appears there is no navigation to display in breadcrumb but the breadcrumb parent element still exists. This checks the last child is set before changing its style.

Before that, this triggered a Javascript error and stopped page scripts. For exemple the notification page did not fetch notifications and it seemed this page was always empty and did not work.