rework: replace initiliaze_call service by inlined creationjoin action in bbbpad_videocall controller
The Ajax request to initiliaze_call and subsequent window.open in response handler asks user for popup authorization.

Since this can prevent users to access a videocall,
we define a simpler workflow where everything happens in the controller.
This means we can't have any feedback in the origin window,
since we open a new window for the videocall.