fix: user regains access to his forum reports through profile pages.
We customize page heading for all activity pages, based on /mod/ prefix.

However, /mod/forum/user.php is not an activity but more of a report page, since it reports all posts and messages made by user in forums.

It can be accessed through user profile or course profile (which limits scope to course forums) but in such cases, there is no real course module info.

So we use the default page heading for such reports.